requirements.txt
Flask==2.2.2
Werkzeug==2.2.2
python-dotenv==0.21.0
pymongo==4.3.3
certifi==2024.2.2